Introduce MeeseeksWorkerFactory
